+++ Changelog.md
@@ -5,6 +5,7 @@ All significant changes are listed in this file.
 ## [1.1.4] - 2022-10-24
 
 - Disable update actress command when scanner is in use
+- Fix exception when cancelling actress update scan
 
 ## [1.1.3] - 2022-10-23